requirements/test.txt
# PIP Requirements (for automated testing with tox)
argparse==1.2.1
django-debug-toolbar==0.11.0
django-devserver==0.7.0
django-extensions==1.2.5
django-hotrunner==0.2.2
django-setuptest==0.1.4
django-sortedm2m==0.6.0
South==0.8.4
